Fixed `pbcodec.Log#BlockIndex` that is wrong when fetched from `transactionTraces.calls`
It appears that our deep mind code that build the actual block index is a bit wrong as it does not consider call that are reverted and blindly increment a counter each time a log is emitted in the chain. After further validation, we have validated that the block index is incremented from top to bottom only for call that were successfuly recorded to the chain state. This means our current `BlockIndex` value is wrong in some blocks.

To add to that, it appears also that `pbcodec.Log` in `TransactionReceipt` didn't had the `BlockIndex` properly populated at all leading all of them to be 0.

To overcome both issues, when decoding the block, we now properly populated both fields with the appropriate rules and ensures consistency with the chain.