Reduced `rpcCalls` allocations (and slightly improve speed)
The `rpc` package found in `eth-go` knowns how to serialize `[]byte` to correct format accepted by Ethereum JSON-RPC so there is no need to do a `Pretty` call on the address, they can be passed as-is. Added tests to prove it.

Refactored `Cache` to be an interface now and `StoreBackedCache` and `NoOpCache` now exists.

Also bump `types` dependency which brought up rename of `MethodSignature` to `Data`.