Added logic to use finalized block from The Merge
The `LIBnum` is not part of the `pbeth.Block` anymore as it's not determistic anymore once the network has switched to Proof of Stake. As such, we have now two computation for the `LIBNum` and the only place where the data can be retrieved now is in `*bstream.Block` enveloppe.