fix polygon handling of 'bor state sync' trx
The firehose instrumentation on polygon (bor) node currently outputs the
state sync 'calls' inside wrapped transactions (from=(nullAddress),
to=0x0000000000000000000000000000000000001001)

The NormalizeBlockInPlace is called in consoleReader and will now take
these special transactions and merge them down in a single transaction
with a deterministic hash based on blocknum + blockhash, aligned with
the behavior of the bor node. NormalizeBlockInPlace only apply these
types of changes when the 'variant' is detected to be `polygon`.

The polygon firehose blocks already require a full reprocessing of the chain at
this point because in some cases, blocks were produced without the
statesync, but we also added the `--variant` flag to the
`upgrade-merged-blocks` command to allow these tweaks.

console_reader now accepts 'FIRE CANCEL_BLOCK' instruction