fix for 'system transactions to for end of validator span
every 6400 blocks, a system transaction is made that does not appear on
the RPC. From what we've seen, it does not contain any logs, but it does
contain state changes.

With this fix, we keep it in firehose but ensure it appears at the end of the block,
with valid index.