Add Optimism poller command and refactor polling functions
The commit refactors functions in 'fireeth' command-line interface (CLI) to adapt to changes in the 'firehose-core' (firehose dependency). This includes addition of an Optimism block poller command that triggers polling from the Optimism RPC , thus leading to the fetching and handling of Ethereum blocks. Smaller changes include modifications in 'go.mod' and 'go.sum' to reflect updated and/or replaced dependencies.